feat: AI websocket接口初步开发

+++ src/main/java/com/juzipi/springbootinit/config/AiConfig.java
@@ -21,5 +21,5 @@ public class AiConfig {
 
     private String apiAddress;
 
-    private Integer temperature;
+    private Double temperature;
 }
